// src/pages/bookv2/hooks/useCustomBookingFields.generated.ts
export type Booking = {
  billingInfo?: InputMaybe<BillingInfo>;
  cancellationPolicy?: InputMaybe<CancellationPolicy>;
  cancellations?: InputMaybe<Array<InputMaybe<Cancellation>>>;
  clientReference?: InputMaybe<Scalars['String']>;
  confirmed?: InputMaybe<Scalars['DateTime']>;
  email: Scalars['String'];
  externalResources?: InputMaybe<ExternalResources>;
  firstName: Scalars['String'];
  flowId?: InputMaybe<Scalars['Int']>;
  id?: InputMaybe<Scalars['Int']>;
  invoice?: InputMaybe<Invoice>;
  lastName: Scalars['String'];
  livnReference?: InputMaybe<Scalars['String']>;
  partyEmailAddress?: InputMaybe<Scalars['String']>;
  partyName?: InputMaybe<Scalars['String']>;
  passThroughReference?: InputMaybe<Scalars['String']>;
  phone: Scalars['String'];
  supplierReference?: InputMaybe<Scalars['String']>;
  tickets?: InputMaybe<Array<InputMaybe<Ticket>>>;
};
// ...
export enum CustomBookingFieldType {
  Boolean = 'BOOLEAN',
  Date = 'DATE',
  Map = 'MAP',
  Number = 'NUMBER',
  Phone = 'PHONE',
  String = 'STRING'
}
// ...
export enum VisibilityLevel {
  Booking = 'BOOKING',
  Both = 'BOTH',
  None = 'NONE',
  Participant = 'PARTICIPANT'
}

export enum VoucherStatus {
  Active = 'ACTIVE',
  OrderCanceled = 'ORDER_CANCELED',
  Redeemed = 'REDEEMED',
  Scheduled = 'SCHEDULED'
}

export type ZoneOffset = {
  totalSeconds?: InputMaybe<Scalars['Int']>;
};

export type ZoneOffsetPayload = {
  totalSeconds?: Maybe<Scalars['Int']>;
};

export type CustomBookingFieldFragment = { fieldType: Types.CustomBookingFieldType, fieldTypeDataForeignId?: number | null, id: number, placeholder?: string | null, productId: string, label?: string | null, publishedLevel: Types.VisibilityLevel, requiredLevel: Types.VisibilityLevel };

export type CustomBookingFieldStringTypeConfigFragment = { id: number, maxLength?: number | null, regex?: string | null, shortDescription: string };

export type CustomBookingFieldMapTypeConfigFragment = { groupId: number, id: number, label?: string | null, resSystemAnswerId?: string | null };